VulkanPipelineCache: fix LRU behavior for layouts.
The "referenceCount" that was added for layouts did not make much sense,
and was never incremented beyond 1. Replaced this with an "age" that
gets incremented each time a command buffer is flushed, and set to 0
every time a layout usage is is inserted into the current command
buffer. This approach is similar to what we were already doing for
pipelines.